Restructure data model files
