+++ lib/util/data_sync.dart
@@ -4,9 +4,9 @@ import 'dart:io';
 
 import 'package:crypto/crypto.dart';
 import 'package:discover_deep_cove/data/db.dart';
-import 'package:discover_deep_cove/data/models/entry_media_pivot.dart';
-import 'package:discover_deep_cove/data/models/fact_file_category.dart';
-import 'package:discover_deep_cove/data/models/fact_file_entry.dart';
+import 'package:discover_deep_cove/data/models/factfile/entry_media_pivot.dart';
+import 'package:discover_deep_cove/data/models/factfile/fact_file_category.dart';
+import 'package:discover_deep_cove/data/models/factfile/fact_file_entry.dart';
 import 'package:discover_deep_cove/data/models/media_file.dart';
 import 'package:discover_deep_cove/env.dart';
 import 'package:discover_deep_cove/util/permissions.dart';
